Lowercase: "invaliddatemessage"
React threw an error:

`Warning: React does not recognize the `invalidDateMessage` prop on a DOM element. If you intentionally want it to appear in the DOM as a custom attribute, spell it as lowercase `invaliddatemessage` instead. If you accidentally passed it from a parent component, remove it from the DOM element.
    in div (created by FormControl)
    in FormControl (created by withStyles(FormControl))
    in withStyles(FormControl) (created by TextField)
    in TextField (created by DateTextField)
    in DateTextField (created by ModalWrapper)
    in div (created by ModalWrapper)
    in ModalWrapper (created by DatePickerWrapper)
    in DatePickerWrapper (created by class_1)
    in form (created by class_1)
    ...
    `