Build umd bundle
Ref https://github.com/dmtrKovalenko/material-ui-pickers/issues/503

In this diff I added umd bundle. Dependencies was slightly shaked
to simplify distribution and consuming.

`classnames` is too small project to be peer dependency. As a regular
one it will introduce less noise.

material-ui is published as stable and there is not need to keep it as
peer. Moreover it's not possible to keep it as external package in umd
because of path fragmentation.

Removed codecov and enzyme from direct dependencies.

This package will be available in `window` object under `material-ui-pickers`
namespace similar to `material-ui` one.

`keycode` is added explicitly as a direct dependency.